Update example_cpFromTablo
just some subjective changes in how I wanted to handle my daily surlatablo task.  I didn't want it to run if a parity check was in progress, in an attempt to speed up the parity check.  This is pretty specific to my docker environment being implemented on an uNRAID system, but oh well.
Also, I broke the query into two chunks:
1) The first queries on original_air_date, which should only match TV episodes.  These will be processed with commercials left in place, but subtitles saved as external SRT files.
2) The second query relies on not setting the clobber override, which should mean that you just pulled all the tv episodes and the only thing remaining is movies.  This second query will zap commercials.